+++ backend/app/Config/Routes.php
@@ -6,3 +6,6 @@ use CodeIgniter\Router\RouteCollection;
  * @var RouteCollection $routes
  */
 $routes->get('/', 'Home::index');
+
+$routes->get('courses/search', 'Courses::search');
+$routes->resource('courses', ['controller' => 'Courses']);