+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial backend
 - Database tables creation and seeding
 - Database models
+- Courses search controller
 
 ## 1.0.0